# backoffice/bff/src/auth/dto.ts
import { IsEmail, IsString, MinLength } from 'class-validator';

export class LoginDto {
  @IsEmail()
  email!: string;

  @IsString()
  @MinLength(1)
  password!: string;
}

export class StepUpDto {
  @IsString()
  @MinLength(1)
  password!: string;
}