# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/signup": {
      "post": {
        "operationId": "AuthController_signup_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "UC-04 · Crear cuenta y obtener token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "summary": "UC-04 · Iniciar sesión",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/step-up": {
      "post": {
        "operationId": "AuthController_stepUp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "NFR-33 · Step-up: reingresar contraseña para acciones sensibles",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/admin/caregivers/pending": {
      "get": {
        "operationId": "AdminCaregiverController_pending_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaregiverResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-19 · Cola de cuidadores pendientes de aprobación",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers": {
      "get": {
        "operationId": "AdminCaregiverController_list_v1",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "approved",
                "rejected",
                "deactivated"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Busca por nombre o zona",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar cuidadores con filtro por estado (paginado)",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}": {
      "get": {
        "operationId": "AdminCaregiverController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-19 · Detalle del cuidador (con documentación para verificar)",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}/approve": {
      "post": {
        "operationId": "AdminCaregiverController_approve_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-19 · Aprobar cuenta (la publica en el marketplace)",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}/reject": {
      "post": {
        "operationId": "AdminCaregiverController_reject_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectCaregiverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-19 · Rechazar cuenta (con motivo)",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}/badges": {
      "put": {
        "operationId": "AdminCaregiverController_badges_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBadgesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-19 · Actualizar insignias de verificación (niveles independientes)",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}/deactivate": {
      "post": {
        "operationId": "AdminCaregiverController_deactivate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeactivateCaregiverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "OQ-8/NFR-31 · Desactivar cuidador (requiere step-up); dispara el ripple encolado a Hiring",
        "tags": [
          "Back-office"
        ]
      }
    },
    "/api/v1/admin/caregivers/{id}/reactivate": {
      "post": {
        "operationId": "AdminCaregiverController_reactivate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "OQ-8 · Reactivar cuidador desactivado (vuelve a visible)",
        "tags": [
          "Back-office"
        ]
      }
    },
...
    "/api/v1/marketplace/caregivers/{id}": {
      "get": {
        "operationId": "MarketplaceController_profile_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaregiverProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-07 · Ver perfil de cuidador",
        "tags": [
          "Marketplace"
        ]
      }
    },
...
  },
  "info": {
    "title": "Keru API",
    "description": "Backend del MVP de Keru (marketplace de cuidadores). Contrato para clientes (app móvil y web). Ver casos de uso en Keru-Casos-de-Uso-MVP.md y reglas en constitution.md.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "account": {
        "type": "apiKey",
        "in": "header",
        "name": "x-account-id"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "familiar@test.com"
          },
          "password": {
            "type": "string",
            "example": "S3gura!123",
            "minLength": 8
          },
          "role": {
            "type": "string",
            "enum": [
              "patient",
              "family",
              "caregiver"
            ],
            "example": "family"
          },
          "displayName": {
            "type": "string",
            "example": "Juan Díaz"
          }
        },
        "required": [
          "email",
          "password",
          "role",
          "displayName"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "patient",
              "family",
              "caregiver",
              "admin"
            ]
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "accountId",
          "email",
          "role",
          "displayName"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "familiar@test.com"
          },
          "password": {
            "type": "string",
            "example": "S3gura!123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "StepUpDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "S3gura!123"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "CaregiverResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "deactivated"
            ]
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "badges": {
            "type": "object",
            "example": {
              "certifications": true,
              "identity": false,
              "background": false
            }
          },
          "rejectionReason": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "displayName",
          "status",
          "specialties",
          "zone",
          "modalities",
          "badges"
        ]
      },
      "CaregiverDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "deactivated"
            ]
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "rates": {
            "type": "object"
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "badges": {
            "type": "object"
          },
          "rejectionReason": {
            "type": "object"
          },
          "reviewedBy": {
            "type": "object"
          },
          "reviewedAt": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "displayName",
          "status",
          "specialties",
          "certifications",
          "availability",
          "rates",
          "zone",
          "modalities",
          "badges",
          "createdAt"
        ]
      },
      "RejectCaregiverDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Certificación de RCP ilegible"
          }
        },
        "required": [
          "reason"
        ]
      },
      "SetBadgesDto": {
        "type": "object",
        "properties": {
          "certifications": {
            "type": "boolean",
            "example": true
          },
          "identity": {
            "type": "boolean",
            "example": true
          },
          "background": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "DeactivateCaregiverDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Denuncia de un familiar"
          }
        }
      },
...
      "CaregiverProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "specialties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "zone": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ratePerHour": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "badges": {
            "type": "object"
          },
          "isFavorite": {
            "type": "boolean",
            "description": "true si está en los favoritos de la cuenta"
          },
          "certifications": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "availability": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "specialties",
          "zone",
          "modalities",
          "ratePerHour",
          "currency",
          "badges",
          "certifications",
          "availability"
        ]
      },
...
    }
  }
}